+++ src/jmetalcpp/metaheuristics/singleObjective/particleSwarmOptimization/StandardPSO2007.cc
@@ -20,7 +20,7 @@
 //  along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 
-#include <StandardPSO2007.hh>
+#include "StandardPSO2007.hh"
 
 
 /**
@@ -216,7 +216,7 @@ SolutionSet * StandardPSO2007::execute() {
 
   //cout << "neighborhood_i " << neighborhood_->getNeighbors(0) << endl;
   cout << "[";
-  for (int i = 0; i < neighborhood_->getNeighbors(0).size()-1; i++) {
+  for (int i = 0; i < (int)neighborhood_->getNeighbors(0).size()-1; i++) {
     cout << neighborhood_->getNeighbors(0).at(i) << ",";
   }
   cout << neighborhood_->getNeighbors(0).back() << "]" << endl;
